hbaapi_build: Correct HBA_PORTSPEED_NOT_NEGOTIATED value in hbaapi.h
The HBA API Port Attributes like PortSpeed and PortSupportedSpeed
are expected to be bitmasks.
But, if for example a given port's supported speeds are 1 Gbit and
10 Gbit then the value for the PortSupportedSpeed comes out as "5"
that is currently being used for HBA_PORTSPEED_NOT_NEGOTIATED.

As per the common HBA API Specification Phase 2 the value for the
HBA_PORTSPEED_NOT_NEGOTIATED is defined as 1<<15.

This patch is to update the HBA_PORTSPEED_NOT_NEGOTIATED value as
per specification.

The hbaapi2.2.patch file before this change only had changes for
HBAAPILIB.c file. But, with the current patch the hbaapi2.2.patch
file has changes for an additional file in the project.

Hence, had to make changes in Makefile to let the hbaapi2.2.patch
applied during build for all files instead of one and removed
while cleanup.

Signed-off-by: Neerav Parikh <Neerav.Parikh@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>