hbaapi: Include the libHBAAPI code from SourceForge
The hbaapi project on SourceForge is currently unmaintained.
This patch copies the code from that project into this
repository. It also removes the copying step from the
INSTALL document.

This patch does not copy over the hba.reg file as it's
not related to Linux. It also does not copy over
readme.txt, makefile.nt or makefile.unix as these files
are not relevant to this new code base.

Signed-off-by: Robert Love <robert.w.love@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>