hbaapi: Apply hbaapi2.2.patch and remove it from the repository
The hbaapi2.2.patch is the accumulation of changes that were
implemented against the SourceForge hbaapi project. A previous
patch added the hbaapi source code to this repository. This patch
applies the hbaapi2.2.patch, removes it from the repository and
updates the Makefile so that no patches are applied durring a build.

Signed-off-by: Robert Love <robert.w.love@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>