INSTALL: Remove unnecessary build step
Now that the SourceForge HBAAPI code has merged
into hbaapi_build to create libHBAAPI we no longer
need to instruct the user to download the code.

This patch updates the INSTALL document to remove
the download step. The patch also updates the header
at the top of the file to remove a reference to
the SourceForge code.

Signed-off-by: Robert Love <robert.w.love@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>